fixing some problems
