+++ public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=af9fb1893bf11a7398f1f441f888b9d0",
+    "/js/app.js": "/js/app.js?id=6bd72e4c4081d0c47fa459b9b0ec34f5",
     "/css/app.css": "/css/app.css?id=7a50e2f8bed5b16d6e3d293318fdf98c"
 }
